refactor: smelly wallet codes
